repmgr: add repl_nodes table to 'repmgr' replication set.
We should probably do the same for repl_events.

Todo: when registering check that the node is in the 'repmgr' replication set;
if not refuse to join. We should provide an optional facility to add
'repmgr' to the node's replication set, but not do this by default
as it involves a node-level BDR configuration change.