repmgrd: initial support for monitoring of BDR nodes by priority
This enables repmgrd to monitor the BDR's "designated master"
(as defined by the node's priority); in the event of that node
failing, the node with the next highest priority will be selected
and will be monitored by all repmgrd instances, including the
one on the failed node.

repmgrd on the failed node will still run as long as active nodes
are available, so each time failure of the designated master is
detected, it can issue event notification commands to e.g. reconfigure
a local connection pooler to point to the new "designated master".